requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
aiosqlite>=0.19.0
pywinpty>=2.0.12
python-dotenv>=1.0.0
httpx>=0.26.0
